make the new caching the default
